Refactor into ingest and search parent components; more styling
