+++ public/pages/workflow_detail/left_nav/nav_content/index.ts
@@ -3,4 +3,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './nav_content';
+export * from './ingest_content';
+export * from './search_content';
