+++ public/pages/workflow_detail/workspace/workspace-styles.scss
@@ -3,5 +3,5 @@
 }
 
 .left-nav-static-width {
-  width: 500px;
+  width: 600px;
 }
